learnlangchain/pyproject.toml
[project]
name = "learnlangchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.5",
    "python-dotenv>=1.2.1",
]
